feat: basic support for windows, no port info
