+++ Cargo.toml
@@ -10,7 +10,7 @@ license-file = "LICENSE"
 
 [dependencies]
 inquire = "0.4.0"
-sysinfo = "0.3.15"
+sysinfo = "0.26.4"
 regex = "1.6.0"
 once_cell = "1"
 unicode-width = "0.1.10"
